+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 -   `<PropertyName />`
     -   provide `labelProps` to configure the automatically injected `Label` element when `PropertyName` is only a string
+-   `<CodeMirror />`
+    -   support for additional modes: jinja2, yaml and json
 
 ### Fixed